refactor(dashboard): extract goals domain (A1 phase 2 final)
Extract goals projections (goal tree + detail + verification decoders
and the two fetchers) into dashboard-goals.ts. With this, dashboard.ts
becomes a pure re-export facade: 4243 -> 313 lines (92.6% reduction),
no domain logic remains.

Goals is structurally distinct from the prior 17 domains: it defines
zero local types (all 17 goal types originate from ../types). So the
facade re-exports those types from ../types (single source, matching
the existing line 709-720 precedent) and only the two fetchers from
./dashboard-goals. This sidesteps the duplicate-identifier / dual-path
hazard: each symbol has exactly one source.

- New: dashboard/src/api/dashboard-goals.ts (550 lines; 523-line body
  extracted verbatim via sed, no manual transcription)
- dashboard.ts: 838 -> 313 lines (goals body removed, header pruned of
  goals-only imports: normalize helpers, get, keeper-store-normalize)
- typecheck green (tsc --noEmit)
- dashboard.test.ts 59/59 passed
- consumer code unchanged: facade re-export preserves all symbols
  (`from './api/dashboard'`)
